refactor(web): extract scheduler page context construction (#654)
First slice of atomizing src/web/routes/scheduler.py: move the heavy context
construction out of the route module into src/web/scheduler/context.py —
collector-health context builder + its presentation helpers, jobs context,
worker-health probe, pipeline-result metadata, notification snapshot payload,
and the JOB_LABELS / heartbeat constants.

routes/scheduler.py re-exports the moved names (including _is_worker_alive,
which src/web/search imports, and the pure helpers + monkeypatch target the
test suite reaches via this module path) so no importer or test changes are
needed. Endpoint orchestration, redirect helpers, and page rendering stay in
the router for the follow-up slice.

Behavior unchanged: same /scheduler/* routes, context keys, worker-health
banner, and pause/resume flow.
