+++ src/web/scheduler/__init__.py
@@ -1,0 +1,1 @@
+"""Scheduler web domain: context builders, handlers, responses."""
